refactor fileStructure
