+++ apps/client/src/app/app.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
-
 import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { AuthModule } from 'libs/auth/src/lib/auth.module'
-import { TodoService } from './todo.service';
+import { TodoService } from '../app/services/todo.service';
+import { CoffeeService } from '../app/services/coffee.service';
 
 @Module({
   imports: [AuthModule],
   controllers: [AppController],
-  providers: [AppService, TodoService],
+  providers: [CoffeeService, TodoService],
 })
 export class AppModule {}
